+++ src/components/modal/modal.js
@@ -2,6 +2,7 @@ import $form from "../form/form.js";
 import $restaurantItem from "../restaurant/restaurant-item.js";
 import { CATEGORY_ICON } from "../../constants/iconPath.js";
 import { FORM_EVENT } from "../form/formEvent.js";
+import $createRestaurantForm from "../form/createRestaurantForm.js";
 
 export const handleModalClose = () => {
   document.querySelector(".modal").classList.remove("modal--open");
@@ -13,11 +14,11 @@ const handleModalCloseEsc = (e) => {
   }
 };
 
-const handleModalOpen = () => {
+export const handleModalOpen = () => {
   document.querySelector(".modal").classList.add("modal--open");
 };
 
-const $modal = (form) => {
+const $modal = (type) => {
   const wrapper = document.createElement("div");
   wrapper.classList.add("modal");
 
@@ -25,22 +26,17 @@ const $modal = (form) => {
   background.classList.add("modal-backdrop");
   wrapper.appendChild(background);
 
-  const container = document.createElement("div");
-  container.classList.add("modal-container");
-
-  const title = document.createElement("h2");
-  title.classList.add("modal-title", "text-title");
-  title.textContent = "새로운 음식점";
-  container.appendChild(title);
-  container.appendChild($form(form, FORM_EVENT.addRestaurant));
-  wrapper.appendChild(container);
+  switch (type) {
+    case "restaurantForm":
+      wrapper.appendChild($createRestaurantForm());
+      break;
+    case "restaurantInfo":
+      wrapper.appendChild();
+      break;
+  }
 
   document.addEventListener("keydown", handleModalCloseEsc);
   background.addEventListener("click", handleModalClose);
-
-  document
-    .querySelector(".gnb__button")
-    .addEventListener("click", handleModalOpen);
 
   return wrapper;
 };
